==> 00_eda.ipynb <==
import functions 

import pandas as pd
import plotly.express as px
import plotly.graph_objects as go


# Reads a text file called 'antenas.txt' with space as delimiter and no header
df = pd.read_csv(   "files\\antenas.txt", 
                    sep = ' ', 
                    header = None)  

# Removes index columns 2 and 6 from the DataFrame
df.drop(columns = [2, 6], 
        inplace = True)

# Rename DataFrame columns to more descriptive names
df.rename(  columns = {0 : "ID", 1 : "time_In", 3 : 'X', 4 : 'Y', 5 : "Permanence (s)"}, 
            inplace = True)

# Multiplies the values in the "Permanence (s)" column by 10
df["Permanence (s)"] = df["Permanence (s)"] * 10

# Display the first rows of the DataFrame
df.head()

# Save the new df in parquet
df.to_parquet("files\\antennas.parquet")

### Exploratory Data Analysis

total_vehicles = df["ID"].nunique()
print(f"Existem {total_vehicles} no cenário.\n\n")


time_df = df[["ID", "Permanence (s)"]].groupby("ID").sum().copy()
time_df.reset_index(inplace = True)

print(f"Informações básicas sobre o tempo de permanência dos veículos.")
time_df["Permanence (s)"].describe()

df_ = df[['X', 'Y', "ID"]].groupby(['X', 'Y']).count()
df_.reset_index(inplace = True)
df_ = df_.pivot(index='X', columns='Y', values="ID")
df_.fillna(0, inplace = True)

fig = px.imshow(df_)
fig.update_layout(  title = "Total de Carros por Coordenadas (X,Y)")

fig.update_xaxes(   title_text = "Coordenada: X")
fig.update_yaxes(   title_text = "Coordenada: Y")

fig.update_xaxes(   tickvals = [0, 20, 40, 60, 80, 99], 
                    ticktext = [0, 20, 40, 60, 80, 99])

fig.update_yaxes(   tickvals = [0, 20, 40, 60, 80, 99], 
                    ticktext = [0, 20, 40, 60, 80, 99])

fig.show()

df_ = df[['X', 'Y', "Permanence (s)"]].groupby(['X', 'Y']).sum()
df_.reset_index(inplace = True)
df_ = df_.pivot(index='X', columns='Y', values="Permanence (s)")
df_.fillna(0, inplace = True)

fig = px.imshow(df_)
fig.update_layout(  title = " Total de Permanência dos Veículos nas Coordanadas  (X,Y)")

fig.update_xaxes(   title_text = "Coordenadas: X")
fig.update_yaxes(   title_text = "Coordenadas: Y")

fig.update_xaxes(   tickvals = [0, 20, 40, 60, 80, 99], 
                    ticktext = [0, 20, 40, 60, 80, 99])

fig.update_yaxes(   tickvals = [0, 20, 40, 60, 80, 99], 
                    ticktext = [0, 20, 40, 60, 80, 99])

fig.show()
